+++ wp-content/themes/core-wp/patterns/content-none.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * Title: No Results
+ * Slug: core-wp/content-none
+ * Categories: text
+ */
+?>
+<!-- wp:group {"layout":{"type":"constrained"}} -->
+<div class="wp-block-group">
+	<!-- wp:heading {"level":2} -->
+	<h2>Nothing Found</h2>
+	<!-- /wp:heading -->
+
+	<!-- wp:paragraph -->
+	<p>It seems we can&rsquo;t find what you&rsquo;re looking for. Perhaps searching can help.</p>
+	<!-- /wp:paragraph -->
+
+	<!-- wp:search {"label":"Search","buttonText":"Search"} /-->
+</div>
+<!-- /wp:group -->