+++ wp-content/themes/core-wp/patterns/content-page.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * Title: Page Content
+ * Slug: core-wp/content-page
+ * Categories: pages
+ * Block Types: core/post-template
+ */
+?>
+<!-- wp:group {"tagName":"article","layout":{"type":"constrained"}} -->
+<article class="wp-block-group">
+	<!-- wp:post-title {"level":1} /-->
+	<!-- wp:post-featured-image /-->
+	<!-- wp:post-content /-->
+</article>
+<!-- /wp:group -->