+++ wp-content/themes/core-wp/patterns/content-post.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * Title: Post Content
+ * Slug: core-wp/content-post
+ * Categories: posts
+ * Block Types: core/post-template
+ */
+?>
+<!-- wp:group {"tagName":"article","layout":{"type":"constrained"}} -->
+<article class="wp-block-group">
+	<!-- wp:post-title {"isLink":true,"level":2} /-->
+	<!-- wp:post-date /-->
+	<!-- wp:post-featured-image {"isLink":true} /-->
+	<!-- wp:post-excerpt /-->
+</article>
+<!-- /wp:group -->