+++ wp-content/themes/core-wp/patterns/content-search.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * Title: Search Result
+ * Slug: core-wp/content-search
+ * Categories: posts
+ * Block Types: core/post-template
+ */
+?>
+<!-- wp:group {"tagName":"article","layout":{"type":"constrained"}} -->
+<article class="wp-block-group">
+	<!-- wp:post-title {"isLink":true,"level":2} /-->
+	<!-- wp:post-date /-->
+	<!-- wp:post-excerpt /-->
+</article>
+<!-- /wp:group -->